collector/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-multipart>=0.0.9
pydantic>=2.6
pyyaml>=6.0
numpy>=1.24
pillow>=10.0
matplotlib>=3.8
